split power

+++ src/rust/src/combination_choose.rs
@@ -1,8 +1,10 @@
-use crate::{choose::Choose, combination::Combination, ops::MulInv};
+use crate::{choose::Choose, combination::Combination};
 
 impl<T> Choose<T> for Combination<T>
 where
     T: std::ops::Mul<Output = T> + From<u64> + Clone,
 {
-    fn choose(&mut self, n: u64, k: u64) -> T { self.calc(n, k) }
+    fn choose(&mut self, n: u64, k: u64) -> T {
+        self.calc(n as usize, k as usize)
+    }
 }
